Implement remove dynamic field logic in base classes instead of trait
As requested in PR #82
