+++ classes/ValidFormBuilder/Password.php
@@ -56,14 +56,6 @@ namespace ValidFormBuilder;
  */
 class Password extends Element
 {
-    use CanRemoveDynamicFields;
-
-    public function __construct($name, $type, $label = "", $validationRules = array(), $errorHandlers = array(), $meta = array())
-    {
-        parent::__construct($name, $type, $label, $validationRules, $errorHandlers, $meta);
-        $this->initialiseDynamicRemoveMeta();
-    }
-
     /**
      * @internal
      * @see \ValidFormBuilder\Element::toHtml()
@@ -108,7 +100,7 @@ class Password extends Element
             }
 
             //*** Add data-dynamic="original" or data-dynamic="clone" attributes to dynamic fields
-            if ($this->getDynamicCount() > 0) {
+            if ($this->isDynamic()) {
                 if ($intCount === 0) {
                     // This is the first, original element. Make sure to define that.
                     $this->setMeta('data-dynamic', 'original', true);
@@ -161,7 +153,7 @@ class Password extends Element
             }
 
             //*** Add data-dynamic="original" or data-dynamic="clone" attributes to dynamic fields
-            if ($this->getDynamicCount() > 0) {
+            if ($this->isDynamic()) {
                 if ($intCount === 0) {
                     // This is the first, original element. Make sure to define that.
                     $this->setMeta('data-dynamic', 'original', true);
