+++ classes/ValidFormBuilder/Select.php
@@ -89,8 +89,6 @@ namespace ValidFormBuilder;
  */
 class Select extends Element
 {
-    use CanRemoveDynamicFields;
-
     /**
      * Collection of option elements created for this select element
      * @internal
@@ -126,8 +124,6 @@ class Select extends Element
         if ($this->__options->count() == 0) {
             $this->__parseRanges();
         }
-
-        $this->initialiseDynamicRemoveMeta();
     }
 
     /**
@@ -176,7 +172,7 @@ class Select extends Element
             }
 
             //*** Add data-dynamic="original" or data-dynamic="clone" attributes to dynamic fields
-            if ($this->getDynamicCount() > 0) {
+            if ($this->isDynamic()) {
                 if ($intCount === 0) {
                     // This is the first, original element. Make sure to define that.
                     $this->setMeta('data-dynamic', 'original', true);
@@ -219,7 +215,7 @@ class Select extends Element
             }
 
             //*** Add data-dynamic="original" or data-dynamic="clone" attributes to dynamic fields
-            if ($this->getDynamicCount() > 0) {
+            if ($this->isDynamic()) {
                 if ($intCount === 0) {
                     // This is the first, original element. Make sure to define that.
                     $this->setMeta('data-dynamic', 'original', true);
