+++ classes/ValidFormBuilder/Text.php
@@ -69,14 +69,6 @@ namespace ValidFormBuilder;
  */
 class Text extends Element
 {
-    use CanRemoveDynamicFields;
-
-    public function __construct($name, $type, $label = "", $validationRules = array(), $errorHandlers = array(), $meta = array())
-    {
-        parent::__construct($name, $type, $label, $validationRules, $errorHandlers, $meta);
-        $this->initialiseDynamicRemoveMeta();
-    }
-
     /**
      * @internal
      * @see \ValidFormBuilder\Element::toHtml()
@@ -121,7 +113,7 @@ class Text extends Element
             }
 
             //*** Add data-dynamic="original" or data-dynamic="clone" attributes to dynamic fields
-            if ($this->getDynamicCount() > 0) {
+            if ($this->isDynamic()) {
                 if ($intCount === 0) {
                     // This is the first, original element. Make sure to define that.
                     $this->setMeta('data-dynamic', 'original', true);
@@ -174,7 +166,7 @@ class Text extends Element
             }
 
             //*** Add data-dynamic="original" or data-dynamic="clone" attributes to dynamic fields
-            if ($this->getDynamicCount() > 0) {
+            if ($this->isDynamic()) {
                 if ($intCount === 0) {
                     // This is the first, original element. Make sure to define that.
                     $this->setMeta('data-dynamic', 'original', true);
