+++ classes/ValidFormBuilder/Textarea.php
@@ -52,8 +52,6 @@ namespace ValidFormBuilder;
  */
 class Textarea extends Element
 {
-    use CanRemoveDynamicFields;
-
     /**
      * Create new Textarea object
      * @internal
@@ -70,8 +68,6 @@ class Textarea extends Element
         }
 
         parent::__construct($name, $type, $label, $validationRules, $errorHandlers, $meta);
-
-        $this->initialiseDynamicRemoveMeta();
     }
 
     /**
@@ -122,7 +118,7 @@ class Textarea extends Element
             }
 
             //*** Add data-dynamic="original" or data-dynamic="clone" attributes to dynamic fields
-            if ($this->getDynamicCount() > 0) {
+            if ($this->isDynamic()) {
                 if ($intCount === 0) {
                     // This is the first, original element. Make sure to define that.
                     $this->setMeta('data-dynamic', 'original', true);
@@ -172,7 +168,7 @@ class Textarea extends Element
             }
 
             //*** Add data-dynamic="original" or data-dynamic="clone" attributes to dynamic fields
-            if ($this->getDynamicCount() > 0) {
+            if ($this->isDynamic()) {
                 if ($intCount === 0) {
                     // This is the first, original element. Make sure to define that.
                     $this->setMeta('data-dynamic', 'original', true);
